Get other versions’ assets from GitHub
Since June 2022, the CDNjs API only returns assets for the latest
version of a library. To get assets for previous versions, we would have
to make one network request per library and version, which is
prohibitively expensive. Instead, get the assets from GitHub; with the
?recursive parameter to their REST API, we can usually get all the files
in the cdnjs/cdnjs repo for a library, across all versions, with a
single network request. (For a few libraries, the response gets
truncated and we have to make one request per version after all, but
this is rare.)

lib["assets"] is changed from a list into a dict (by version), so that
we can avoid overwriting the CDNjs assets for the latest version with
GitHub’s data; this also means that template.html.j2 can directly
address the latest version instead of looping over all of them. The
dictsort in modal-template.html.j2 isn’t great (29, 3, 30, …), but it’s
better than nothing, and I think it’s the best we have easily available.
(Even the CDNjs API returns versions in this bad order: see for example
<https://api.cdnjs.com/libraries/three.js?fields=versions>.)

The lib["assets"] change causes problems for at least one library,
'xls', where the latest version returned by CDNjs doesn’t actually have
any assets on GitHub (nor on CDNjs, for that matter). Work around that
by not emitting any assets in the HTML template then. (Previously, I
think the condition {%- if asset.version == lib.version -%} would never
have been true, so it would’ve come out to mostly the same thing.) Since
I assume this only happens for very rare libraries that nobody is likely
to care about anyways, just emit an HTML comment for now (so there’s
*some* clue about what happened, but we don’t have to make the design
nice or anything); if anyone complains, we can look into emitting nicer
non-comment HTML in this case (or handling it better otherwise), but for
now it’s not worth it IMHO.

Bug: T342519
Change-Id: I7afdfd57dd79e34c342b7e1b7e5344e346b64c34