Write mod*.html files immediately
Write out each library’s individual HTML file (for the versions) as soon
as we have the data, then throw away all the assets except for the
latest version (which is the only one that index.html uses). This should
save a lot of memory (on Toolforge, a job with 4 GiB of memory was still
killed after just 383 out of 4420 libraries, so I don’t think holding
all the assets of all library versions in memory at once is very
feasible anymore), and also it provides a nice progress indication while
the script is running: just ls the output directory.

Bug: T342519
Change-Id: Ib6081437c8fefb9c37fd001b211de0bf7421be8d