Retry GitHub requests if rate limit is hit
This is unlikely while we’re only looking up stars, but becomes more of
an issue once we start making a lot more requests in order to get
additional library versions; add the needed method separately first.

Also update the request headers while we’re at it, according to GitHub’s
documentation.

The status code is returned to callers, because it feels right to do it,
but currently ignored to match the existing code. (Asserting that
status_code == 200 crashes some libs, such as react-router-dom, where
the API returns 404, probably because we’re not quite using the right
repo name but I can’t be bothered to figure it out.)

Bug: T342519
Change-Id: I3d725abecfd42583eff49b0b2a801d3853cadd3d